Fix bug in scrot detection on Linux
The problem with the original subprocess.Popen call was that with
shell=True the whole command, including arguments, need to go in the
first string, and subsequent strings are fed as arguments to /bin/sh,
which always results in returncode 1.

Since shell=True is not needed anyway, this problem can be solved by
removing it.

Then there's a very hypothetical deadlock problem with the use of the
wait() method, and just to be safe I fix that too by using the
solution discussed here: http://stackoverflow.com/a/5195443/566644

Finally change the except clause to only catch the expected error and
let all other errors propagate, to avoid the error hiding antipattern.