Expose confidence setting in all locate functions
When OpenCV is available, it is used to perform the image matching and
it allows a confidence value to be adjusted.  This setting was only
being exposed through the locateAll function and not the rest of the
functions.  Since the rest of the locate functions call locateAll, this
changes them to pass all keyword arguments through including the
confidence setting.