fix cropped image saving on Linux and Mac
As referenced in pyautogui's issue 286 (https://github.com/asweigart/pyautogui/issues/286), currently a call to the screenshot function that includes both a region and filename will save an image of the entire screen to the filesystem, but return a cropped image. 
This change would make the image saved on the filesystem (for osx and linux) also be the cropped version, which is what the user would expect, and mirrors the behavior on win32.